Improved error handling and implemented logging for the setup class

+++ src/us/stomberg/solarsystemsim/physics/Physics.java
@@ -101,7 +101,7 @@ public class Physics {
                     }
                 }
             }
-            duration += repMillis;
+            duration += (long) repMillis;
             reps--;
         }
     }
